freeboard - file upload delete complete

+++ src/controller/FileDeleteController.java
@@ -7,6 +7,9 @@ import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
+
+import com.google.gson.Gson;
 
 import Service.FileService;
 
@@ -27,13 +30,24 @@ public class FileDeleteController extends HttpServlet {
 	protected void doAction(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
 		String param = request.getParameter("fileNo");
 		
+		boolean result = false;
+		HttpSession session = request.getSession();
+		
 		//파라미터 값 널체크 후 파일 삭제
 		if(param != null && param !="") {
 			int fileNo = Integer.parseInt(param);
-			
-			fsvc.deleteFile(fileNo);
+			//인증처리 - 로그인사용자와 보드 작성자 일치하는지 확인
+			String owner = fsvc.getFileOwner(fileNo);
+			if(owner.equals(session.getAttribute("userId"))) {
+				fsvc.deleteFile(fileNo);
+				result = true;
+			} 
 		}
+		
+		response.getWriter().println(
+				"{\"result\": "+ result +"}");
+		
 		//이전 페이지로 이동
-		response.sendRedirect(request.getHeader("referer"));
+//		response.sendRedirect(request.getHeader("referer"));
 	}
 }
